Progress: Continue fixing the unit tests.
Properly handle mock return types.

Make sure to return F_file_descriptor_not instead of F_stream_not in cases where the file descriptor is checked instead of the file stream.

+++ level_0/f_conversion/tests/unit/c/mock-conversion.c
@@ -14,7 +14,7 @@ size_t __wrap_fwrite_unlocked(const void * const ptr, size_t size, size_t nmemb,
   const bool failure = mock_type(bool);
 
   if (failure) {
-    return mock_type(int);
+    return mock_type(size_t);
   }
 
   const bool check_ptr = mock_type(bool);
