+++ level_0/f_print/tests/unit/c/mock-print.c
@@ -17,7 +17,7 @@ size_t __wrap_fwrite_unlocked(const void *ptr, size_t size, size_t nmemb, FILE *
   const bool failure = mock_type(bool);
 
   if (failure) {
-    return mock_type(int);
+    return mock_type(size_t);
   }
 
   return mock_type(size_t);
@@ -28,7 +28,7 @@ ssize_t __wrap_write(int fd, const void *buf, size_t count) {
   const bool failure = mock_type(bool);
 
   if (failure) {
-    errno = mock_type(int);
+    errno = mock_type(ssize_t);
 
     return -1;
   }
